# web/openapi.json
{
  "components": {
    "schemas": {
      "AdapterStatus": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "gaps": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "harness": {
            "enum": [
              "claude",
              "codex",
              "hermes"
            ],
            "title": "Harness",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "harness",
          "ok"
        ],
        "title": "AdapterStatus",
        "type": "object"
      },
      "CompletionDetail": {
        "properties": {
          "completed_at": {
            "title": "Completed At",
            "type": "number"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "harness": {
            "enum": [
              "claude",
              "codex",
              "hermes"
            ],
            "title": "Harness",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "machine_id",
          "harness",
          "session_id",
          "project_id",
          "title",
          "completed_at",
          "excerpt",
          "response"
        ],
        "title": "CompletionDetail",
        "type": "object"
      },
      "CompletionObs": {
        "description": "A settled response cycle with its delivered final response.",
        "properties": {
          "completed_at": {
            "title": "Completed At",
            "type": "number"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "harness": {
            "enum": [
              "claude",
              "codex",
              "hermes"
            ],
            "title": "Harness",
            "type": "string"
          },
          "project": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "harness",
          "session_id",
          "request_id",
          "completed_at",
          "response"
        ],
        "title": "CompletionObs",
        "type": "object"
      },
      "CompletionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompletionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CompletionPage",
        "type": "object"
      },
      "CompletionSummary": {
        "properties": {
          "completed_at": {
            "title": "Completed At",
            "type": "number"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "harness": {
            "enum": [
              "claude",
              "codex",
              "hermes"
            ],
            "title": "Harness",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "machine_id",
          "harness",
          "session_id",
          "project_id",
          "title",
          "completed_at",
          "excerpt"
        ],
        "title": "CompletionSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectRef": {
        "properties": {
          "checkout": {
            "title": "Checkout",
            "type": "string"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "title": "Keys",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "keys",
          "name",
          "checkout"
        ],
        "title": "ProjectRef",
        "type": "object"
      },
...
      "Report": {
        "description": "One collector upload: a full current snapshot plus queued completions.",
        "properties": {
          "adapters": {
            "items": {
              "$ref": "#/components/schemas/AdapterStatus"
            },
            "title": "Adapters",
            "type": "array"
          },
          "completions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CompletionObs"
            },
            "title": "Completions",
            "type": "array"
          },
          "expired_total": {
            "default": 0,
            "title": "Expired Total",
            "type": "integer"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "machine_name": {
            "title": "Machine Name",
            "type": "string"
          },
          "sent_at": {
            "title": "Sent At",
            "type": "number"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionObs"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "machine_id",
          "machine_name",
          "sent_at",
          "sessions",
          "adapters"
        ],
        "title": "Report",
        "type": "object"
      },
      "ReportAck": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "ReportAck",
        "type": "object"
      },
      "SessionObs": {
        "properties": {
          "activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "gap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          },
          "harness": {
            "enum": [
              "claude",
              "codex",
              "hermes"
            ],
            "title": "Harness",
            "type": "string"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "number"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "project": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "working",
              "waiting",
              "idle",
              "unknown"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "harness",
          "session_id",
          "state",
          "observed_at"
        ],
        "title": "SessionObs",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Crew",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/completions": {
      "get": {
        "operationId": "completions_api_completions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Completions"
      }
    },
    "/api/completions/{completion_id}": {
      "get": {
        "operationId": "completion_api_completions__completion_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "completion_id",
            "required": true,
            "schema": {
              "title": "Completion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Completion"
      }
    },
...
    "/api/ingest": {
      "post": {
        "operationId": "ingest_api_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Report"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest"
      }
    },
...
  }
}